Add CentOS-based 3-master Mesosphere deployment template
This commmit will add a template that deploys a three-master Mesosphere
cluster to a bunch of CentOS machines. This template includes a bunch of
goodies that are useful to devops folks, like by-default SSH login.

It also includes public IP addresses for BOTH the slave and master
nodes. This is for quick, convenient SSH access, but arguably neither
should be public. Instead, a production cluster should run on a private
network and operators should VPN in, at which point they can choose to
SSH into whatever node they want. But, that is a little heavy handed for
our demo templates, hence, in this, everything is public.