Update the Mesosphere default config script commands
Mesosphere nodes are currently configured by downloading a shell script
from GitHub and running on each node. Currently the template is
configured to have a default config script to run, but not a default
command to run it. So users have to manually put in the command `sh
configure_master.sh` each time. This of course should come by default.

This commit will introduce those defaults to those templates.